changes to CR
